save and load layer info, moved layermanager to core
