+++ avogadro/core/molecule.cpp
@@ -9,12 +9,14 @@
 #include "color3f.h"
 #include "cube.h"
 #include "elements.h"
+#include "layermanager.h"
 #include "mesh.h"
 #include "residue.h"
 #include "unitcell.h"
 
 #include <algorithm>
 #include <cassert>
+#include <iostream>
 
 namespace Avogadro {
 namespace Core {
@@ -22,7 +24,8 @@ namespace Core {
 using std::swap;
 
 Molecule::Molecule()
-  : m_basisSet(nullptr), m_unitCell(nullptr), m_graphDirty(false)
+  : m_basisSet(nullptr), m_unitCell(nullptr), m_graphDirty(false),
+    m_layers(LayerManager::getMoleculeLayer(this))
 {}
 
 Molecule::Molecule(const Molecule& other)
@@ -39,7 +42,8 @@ Molecule::Molecule(const Molecule& other)
     m_unitCell(other.m_unitCell ? new UnitCell(*other.m_unitCell) : nullptr),
     m_residues(other.m_residues), m_graph(other.m_graph),
     m_graphDirty(other.m_graphDirty), m_bondPairs(other.m_bondPairs),
-    m_bondOrders(other.m_bondOrders), m_atomicNumbers(other.m_atomicNumbers)
+    m_bondOrders(other.m_bondOrders), m_atomicNumbers(other.m_atomicNumbers),
+    m_layers(LayerManager::getMoleculeLayer(&other, this))
 {
   // Copy over any meshes
   for (Index i = 0; i < other.meshCount(); ++i) {
@@ -73,7 +77,8 @@ Molecule::Molecule(Molecule&& other) noexcept
     m_graphDirty(std::move(other.m_graphDirty)),
     m_bondPairs(std::move(other.m_bondPairs)),
     m_bondOrders(std::move(other.m_bondOrders)),
-    m_atomicNumbers(std::move(other.m_atomicNumbers))
+    m_atomicNumbers(std::move(other.m_atomicNumbers)),
+    m_layers(LayerManager::getMoleculeLayer(&other, this))
 {
   m_basisSet = other.m_basisSet;
   other.m_basisSet = nullptr;
@@ -127,6 +132,8 @@ Molecule& Molecule::operator=(const Molecule& other)
     m_unitCell = other.m_unitCell ? new UnitCell(*other.m_unitCell) : nullptr;
   }
 
+  m_layers = LayerManager::getMoleculeLayer(&other, this);
+
   return *this;
 }
 
@@ -168,15 +175,27 @@ Molecule& Molecule::operator=(Molecule&& other) noexcept
     other.m_unitCell = nullptr;
   }
 
+  m_layers = LayerManager::getMoleculeLayer(&other, this);
+
   return *this;
 }
 
 Molecule::~Molecule()
 {
+  LayerManager::deleteMolecule(this);
   delete m_basisSet;
   delete m_unitCell;
   clearMeshes();
   clearCubes();
+}
+
+Layer& Molecule::layer()
+{
+  return m_layers;
+}
+const Layer& Molecule::layer() const
+{
+  return m_layers;
 }
 
 void Molecule::setData(const std::string& name, const Variant& value)
@@ -275,6 +294,7 @@ Molecule::AtomType Molecule::addAtom(unsigned char number)
     m_graph.addVertex();
   }
   m_atomicNumbers.push_back(number);
+  m_layers.addAtomToActiveLayer(atomCount() - 1);
   return AtomType(this, static_cast<Index>(atomCount() - 1));
 }
 
@@ -329,6 +349,7 @@ void Molecule::swapAtom(Index a, Index b)
       m_graph.addEdge(pair.first, pair.second);
     }
   }
+  m_layers.swapLayer(a, b);
 }
 
 bool Molecule::removeAtom(Index index)
@@ -362,6 +383,7 @@ bool Molecule::removeAtom(Index index)
   }
   // the bonds from back() now are in index, so we need to rebond it
   rebondBond(index, affectedIndex);
+  m_layers.removeAtom(index);
   return true;
   return true;
 }
@@ -1061,6 +1083,18 @@ void Molecule::rebondBond(Index newIndex, Index oldIndex)
     }
   }
   m_graphDirty = true;
+}
+
+std::list<Index> Molecule::getAtomsAtLayer(size_t layer)
+{
+  std::list<Index> result;
+  // get the index in decreasing order so deleting won't corrupt data
+  for (Index i = atomCount(); i > 0; --i) {
+    if (m_layers.getLayerID(i - 1) == layer) {
+      result.push_back(i - 1);
+    }
+  }
+  return result;
 }
 
 } // namespace Core
