+++ avogadro/qtplugins/vanderwaals/vanderwaals.h
@@ -26,12 +26,15 @@ public:
   void process(const QtGui::Molecule& molecule,
                Rendering::GroupNode& node) override;
 
-  QString name() const override { return tr("Van der Waals"); }
+  QString name() const override { return tr(m_name.c_str()); }
 
   QString description() const override
   {
     return tr("Simple display of VdW spheres.");
   }
+
+private:
+  std::string m_name = "Van der Waals";
 };
 } // namespace QtPlugins
 } // namespace Avogadro
